Move Journey class to flow
This is one of several classes that seem to relate to the flow of filling out forms.
